+++ maia/pytree/core/graph/build.py
@@ -1,4 +1,4 @@
-from maia.pytree.graph import algo
+from .algo import depth_first_search
 
 
 class _build_tree_visitor:
@@ -37,7 +37,7 @@ class _build_tree_visitor:
 
 def depth_first_build_trees(g, node_constructor, pre=None):
   v = _build_tree_visitor(node_constructor, pre)
-  algo.depth_first_search(g, v)
+  depth_first_search(g, v)
   return v.retrieve_composition_term()
 
 def depth_first_build(g, node_constructor, pre=None):
@@ -53,7 +53,7 @@ def depth_first_build(g, node_constructor, pre=None):
         `to_children` is a list of the previous sub-graphs of `g_out`
         `to_node` is the new graph node that is to be created.
 
-    pre: tells the algorithm what to do on each node traversed (step.into, step.over, step.out)
+    pre: tells the algorithm what to do on each node traversed (Step.INTO, Step.OVER, Step.OUT)
         More precisely, `pre` is a function of argument `node` and returning a `graph.algo.step` value.
   """
   ts = depth_first_build_trees(g, node_constructor, pre)